Adds a timeout to launchpadlib connections.
The default timeout in LPlib is None which means it will just
hang. Lets set our default to 5 seconds to fix any queries
that hang.

This fixes an issue which started a week ago (when a new
version of LP was released) where reviewday would just hang
when trying to query LP for some of the bug/BP information.

Change-Id: Ia7af4efba543b25d220ccb193abbabc20b9f489f
Reviewed-on: https://review.openstack.org/28315
Approved: Dan Prince <dprince@redhat.com>
Reviewed-by: Dan Prince <dprince@redhat.com>
Tested-by: Jenkins