Make reviewday.json world readable.
Because we were using a python tempfile the reviewday.json
file was readable/writable only by the user that runs
bin/reviewday. Ideally this file would be world readable so
that Apache can host it properly.

In this commit we drop the use of python tempfile in
favor of just plain old open() which respects the
normal file permission settings.

This should fix a Forbidden access error I get when
trying to access http://status.openstack.org/reviews/reviewday.json
today.

Closes-bug: #1273833

Change-Id: I9e9a76e29f5c71f2496eef96a6dbf2496c42d973