Add --project-file option
This patch adds a new -p (--project-file) option which can
be used to specify an external grouping of projects used
to generate the report.

The format of the project file can be YAML or JSON (YAML example
is included in the README.md). The format is a list of dict's to
leave open the option for future expansion (fields other than
'name' might also be useful.)

Change-Id: I5e74fa118296c4531c988f4804749d6a6c1b957a