Generate data_table.html
This patch updates reviewday so that it generates
both an index.html and a separate data_table.html
which can be used for importing the reviewday HTML
table into an externally styled website.

The existing reviewday OpenStack style remains the same.
This patch should prove useful for sites wishing
to re-use the "data table" format as-is inside of a
separately themed web site.

Change-Id: If9a69235484038e4b10ee8a2c66ed834f5f6a8d2